move cllocation array to it's own file in corelocation folder
